Auction announcements
Added map to Frontend struct such that ListenForWinner and SendBid use same stream

Client now doesn't begin listening for winners until they have send their first bid

Switched from using slice to store client streams to using a map as it was easier to check if stream was already inside the map. The adding process was also moved inside the for loop to ensure that it was only when a valid BidRequest was received that we would register users.